+++ packages/apply-release-plan/src/get-changelog-entry.ts
@@ -25,6 +25,7 @@ async function generateChangesForVersionTypeMarkdown(
 
 // release is the package and version we are releasing
 export default async function getChangelogEntry(
+  cwd: string,
   release: ModCompWithPackage,
   releases: ModCompWithPackage[],
   changesets: NewChangesetWithCommit[],
@@ -69,7 +70,13 @@ export default async function getChangelogEntry(
       versionRange &&
       (usesWorkspaceRange || validRange(versionRange) !== null) &&
       shouldUpdateDependencyBasedOnConfig(
-        { type: rel.type, version: rel.newVersion },
+        cwd,
+        {
+          type: rel.type,
+          version: rel.newVersion,
+          oldVersion: rel.oldVersion,
+          dir: rel.dir,
+        },
         {
           depVersionRange: versionRange,
           depType: dependencyVersionRange ? "dependencies" : "peerDependencies",