+++ packages/assemble-release-plan/src/index.ts
@@ -195,6 +195,7 @@ function assembleReleasePlan(
     let dependentAdded = determineDependents({
       releases,
       packagesByName,
+      rootDir: packages.rootDir,
       dependencyGraph,
       preInfo,
       config: refinedConfig,